#505 worldunit.set_deal_tree_mandates
