+++ src/f07_fisc/examples/example_fiscs.py
@@ -2,6 +2,7 @@ from src.f01_road.road import create_road
 from src.f02_bud.healer import healerlink_shop
 from src.f02_bud.item import itemunit_shop
 from src.f02_bud.reason_item import factunit_shop, FactUnit
+from src.f02_bud.bud import budunit_shop, BudUnit
 from src.f05_listen.hubunit import hubunit_shop
 from src.f07_fisc.fisc import FiscUnit, fiscunit_shop
 from src.f07_fisc.examples.fisc_env import get_test_fisc_mstr_dir
@@ -268,3 +269,88 @@ def example_casa_dirty_factunit() -> FactUnit:
     floor_road = create_road(casa_road, "floor status")
     dirty_road = create_road(floor_road, "dirty")
     return factunit_shop(floor_road, dirty_road)
+
+
+def _example_empty_bob_budunit() -> BudUnit:
+    a23_str = "accord23"
+    return budunit_shop("Bob", a23_str)
+
+
+def get_bob_mop_without_reason_budunit_example() -> BudUnit:
+    bob_bud = _example_empty_bob_budunit()
+    casa_str = "casa"
+    floor_str = "floor status"
+    clean_str = "clean"
+    dirty_str = "dirty"
+    mop_str = "mop"
+    casa_road = bob_bud.make_l1_road(casa_str)
+    floor_road = bob_bud.make_road(casa_road, floor_str)
+    clean_road = bob_bud.make_road(floor_road, clean_str)
+    dirty_road = bob_bud.make_road(floor_road, dirty_str)
+    mop_road = bob_bud.make_road(casa_road, mop_str)
+    bob_bud.add_item(casa_road, 1)
+    bob_bud.add_item(floor_road, 1)
+    bob_bud.add_item(clean_road, 1)
+    bob_bud.add_item(dirty_road, 1)
+    bob_bud.add_item(mop_road, 1, pledge=True)
+    return bob_bud
+
+
+def get_bob_mop_with_reason_budunit_example() -> BudUnit:
+    """owner_name: bob, fisc_title: accord23"""
+    bob_bud = get_bob_mop_without_reason_budunit_example()
+    casa_str = "casa"
+    floor_str = "floor status"
+    dirty_str = "dirty"
+    mop_str = "mop"
+    casa_road = bob_bud.make_l1_road(casa_str)
+    floor_road = bob_bud.make_road(casa_road, floor_str)
+    dirty_road = bob_bud.make_road(floor_road, dirty_str)
+    mop_road = bob_bud.make_road(casa_road, mop_str)
+    bob_bud.edit_item_attr(mop_road, reason_base=floor_road, reason_premise=dirty_road)
+    return bob_bud
+
+
+def get_bob_mop_fact_clean_budunit_example() -> BudUnit:
+    bob_bud = get_bob_mop_with_reason_budunit_example()
+    bob_bud.add_acctunit("Bob")
+    casa_road = bob_bud.make_l1_road("casa")
+    floor_road = bob_bud.make_road(casa_road, "floor status")
+    clean_road = bob_bud.make_road(floor_road, "clean")
+    bob_bud.add_fact(floor_road, clean_road)
+    return bob_bud
+
+
+def get_yao_run_with_reason_budunit_example() -> BudUnit:
+    yao_bud = budunit_shop("Yao", "accord23")
+    sport_str = "sport"
+    participate_str = "participate"
+    ski_str = "skiing"
+    run_str = "running"
+    weather_str = "weather"
+    raining_str = "raining"
+    snowing_str = "snowing"
+    sport_road = yao_bud.make_l1_road(sport_str)
+    participate_road = yao_bud.make_road(sport_road, participate_str)
+    ski_road = yao_bud.make_road(participate_road, ski_str)
+    run_road = yao_bud.make_road(participate_road, run_str)
+    weather_road = yao_bud.make_l1_road(weather_str)
+    rain_road = yao_bud.make_road(weather_road, raining_str)
+    snow_road = yao_bud.make_road(weather_road, snowing_str)
+    yao_bud.add_item(participate_road)
+    yao_bud.add_item(ski_road, 5, pledge=True)
+    yao_bud.add_item(run_road, 1, pledge=True)
+    yao_bud.add_item(weather_road)
+    yao_bud.add_item(rain_road)
+    yao_bud.add_item(snow_road)
+    yao_bud.edit_item_attr(ski_road, reason_base=weather_road, reason_premise=snow_road)
+    yao_bud.edit_item_attr(run_road, reason_base=weather_road, reason_premise=rain_road)
+    return yao_bud
+
+
+def get_yao_run_rain_fact_budunit_example() -> BudUnit:
+    yao_bud = get_yao_run_with_reason_budunit_example()
+    weather_road = yao_bud.make_l1_road("weather")
+    rain_road = yao_bud.make_road(weather_road, "raining")
+    yao_bud.add_fact(weather_road, rain_road)
+    return yao_bud
