+++ src/f11_world/world.py
@@ -53,6 +53,7 @@ from src.f10_etl.transformers import (
     etl_create_fisc_deal_trees,
     etl_set_deal_trees_found_facts,
     etl_set_deal_trees_decrees,
+    etl_set_deal_tree_mandates,
 )
 from dataclasses import dataclass
 from sqlite3 import connect as sqlite3_connect, Connection as sqlite3_Connection
@@ -233,6 +234,9 @@ class WorldUnit:
 
     def set_deal_trees_decrees(self):
         etl_set_deal_trees_decrees(self._fisc_mstr_dir)
+
+    def set_deal_tree_mandates(self):
+        etl_set_deal_tree_mandates(self._fisc_mstr_dir)
 
     def mine_to_forecasts(self):  # sourcery skip: extract-method
         fisc_mstr_dir = create_path(self._world_dir, "fisc_mstr")
